chore: rename player file
